requirements.txt
streamlit==1.32.0
openai==1.52.2
python-dotenv==1.0.1
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
beautifulsoup4==4.12.3
langchain==0.1.9
langchain-community==0.0.27 
langchain_text_splitters==0.3.8
nltk==3.9.1
transformers==4.52.4
faiss-cpu==1.11.0